Agregué la pelea entre los heroes y las bestias, modificando a caca uno de los valores numericos de int a double ya que la vida no siempre será positiva

+++ Bestias.java
@@ -1,6 +1,6 @@
 public class Bestias extends Personaje{
     
-    public Bestias(int vida, String nombre, int armadura){
+    public Bestias(double vida, String nombre, double armadura){
         super(vida,nombre,90,armadura);
     }
